refactor: 型定義を適切なファイルに整理
- types.tsを新規作成し、リクエスト型とEsaConfig型を集約
- EsaSearchResultをcaseConverter.tsに移動（変換に関連するため）
- index.tsから全ての型定義を削除し、ビジネスロジックに集中
- 各ファイルのimportを適切に更新
